update test entry point
